Check biala lista
The white list of VAT payers is a list of those indicated as VAT payers, unregistered/deleted/restored to VAT. Check was established by Polish legislation in 2019. This script allow you to automatically check multiple recipients at once. Additional module to be installed is selenium, its webdriver in a class is by default operating on firefox web browser.